:zap::art::hammer: refactor: Update module names & Improve structure
