+++ django_logging/constants/__init__.py
@@ -1,3 +1,3 @@
-from .format_options import FORMAT_OPTIONS
-from .defaults import DefaultLoggingSettings
-from .format_specifiers import LOG_FORMAT_SPECIFIERS
+from .log_format_options import FORMAT_OPTIONS
+from .default_settings import DefaultLoggingSettings
+from .log_format_specifiers import LOG_FORMAT_SPECIFIERS
