+++ django_logging/constants/config_types.py
