+++ django_logging/constants/required_email_settings.py
