+++ django_logging/formatters/__init__.py
@@ -1,1 +1,1 @@
-from .colorized_formatter import ColorizedFormatter
+from .colored_formatter import ColorizedFormatter
