+++ django_logging/formatters/colored_formatter.py
@@ -1,6 +1,6 @@
 import logging
 from django_logging.settings.conf import LogConfig
-from django_logging.utils.colorizer import colorize_log_format
+from django_logging.utils.console_colorizer import colorize_log_format
 
 
 class ColorizedFormatter(logging.Formatter):
