+++ django_logging/handlers/email_handler.py
@@ -3,8 +3,8 @@ import logging
 from django.conf import settings
 from django.template import engines
 from django.utils.timezone import now
-from django_logging.utils.email_notifier import send_email_async
-from django_logging.utils.get_config import use_email_notifier_template
+from django_logging.utils.log_email_notifier import send_email_async
+from django_logging.utils.get_conf import use_email_notifier_template
 from django_logging.middleware import RequestLogMiddleware
